Maps/Warnings: Reduce the amount of warnings/err...

+++ src/tools/ToolsCata/vmap_tools/vmap4_extractor/vmapexport.cpp
@@ -112,7 +112,7 @@ typedef struct
 
 map_id * map_ids;
 uint16 *LiqType = 0;
-uint32 map_count;
+size_t map_count;
 char output_path[128]=".";
 char input_path[1024]=".";
 bool preciseVectorData = false;
